pyproject.toml
[tool.poetry]
name = "noteblock-generator"
version = "0.1.9.2"
description = "Generate music compositions in Minecraft noteblocks"
authors = ["Felix Fourcolor <felix.fourcolor@gmail.com>"]
readme = "README.md"
license = "CC0-1.0"
packages = [{include = "noteblock_generator"}]
repository = "https://github.com/FelixFourcolor/noteblock-generator"
keywords = ["minecraft", "noteblock"]

[tool.poetry.scripts]
noteblock-generator = "noteblock_generator.__main__:main"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
amulet-core = { git = "https://github.com/Amulet-Team/Amulet-Core", branch = "1.0", rev = "db94a3a" }
colorama = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
strictParameterNoneValue = false

[tool.isort]
profile = "black"